add components for color-selector and classification-panel
