+++ geoda-ai/src/components/mapping/mapping-panel.tsx
@@ -1,13 +1,12 @@
 import {useIntl} from 'react-intl';
-import {Select, SelectItem, Button, Tabs, Tab, Card, CardBody, Spacer} from '@nextui-org/react';
-import {useMemo, useState} from 'react';
+import {Button, Tabs, Tab, Card, CardBody, Spacer} from '@nextui-org/react';
+import {useState} from 'react';
 import {useDispatch, useSelector} from 'react-redux';
 import {
   appInjector,
   LayerListFactory,
   makeGetActionCreators,
-  MapManagerFactory,
-  ColorPalette
+  MapManagerFactory
 } from '@kepler.gl/components';
 import {LayerClasses} from '@kepler.gl/layers';
 import {GeoDaState} from '@/store';
@@ -15,80 +14,17 @@ import {MAP_ID, MappingTypes} from '@/constants';
 import {createMapBreaks, createCustomScaleMap} from '@/utils/mapping-functions';
 import {WarningBox, WarningType} from '../common/warning-box';
 import {RightPanelContainer} from '../common/right-panel-template';
-import {getColumnData, getLayer, getNumericFieldNames} from '@/utils/data-utils';
+import {getColumnData, getLayer} from '@/utils/data-utils';
 import {wrapTo} from '@kepler.gl/actions';
-import {SIDEBAR_PANELS, COLOR_RANGES, ColorRange} from '@kepler.gl/constants';
+import {SIDEBAR_PANELS} from '@kepler.gl/constants';
+import {getDefaultColorRange} from '../common/color-selector';
+import {ClassificationPanel, ClassificationOnValuesChange} from '../common/classification-panel';
 
 // const MapContainer = KeplerInjector.get(MapContainerFactory);
 const LayerList = appInjector.get(LayerListFactory);
 const MapManager = appInjector.get(MapManagerFactory);
 
 const NO_MAP_LOADED_MESSAGE = 'Please load a map first before creating and managing your maps.';
-
-export const mappingTypes = [
-  {
-    label: 'Quantile Map',
-    value: MappingTypes.QUANTILE
-  },
-  {
-    label: 'Natural Breaks Map',
-    value: MappingTypes.NATURAL_BREAK
-  },
-  {
-    label: 'Equal Interval Map',
-    value: MappingTypes.EQUAL_INTERVAL
-  },
-  {
-    label: 'Percentile Map',
-    value: MappingTypes.PERCENTILE
-  },
-  {
-    label: 'Box Map (Hinge=1.5)',
-    value: MappingTypes.BOX_MAP_15
-  },
-  {
-    label: 'Box Map (Hinge=3.0)',
-    value: MappingTypes.BOX_MAP_30
-  },
-  {
-    label: 'Standard Deviation Map',
-    value: MappingTypes.STD_MAP
-  },
-  {
-    label: 'Unique Values Map',
-    value: MappingTypes.UNIQUE_VALUES
-  },
-  {
-    label: 'Co-location Map',
-    value: MappingTypes.COLOCATION
-  },
-  {
-    label: 'Raw Rate Map',
-    value: MappingTypes.RAW_RATE
-  },
-  {
-    label: 'Excess Rate Map',
-    value: MappingTypes.EXCESS_RATE
-  },
-  {
-    label: 'Empirical Bayes Map',
-    value: MappingTypes.EB_MAP
-  },
-  {
-    label: 'Spatial Rate Map',
-    value: MappingTypes.SPATIAL_RATE
-  },
-  {
-    label: 'Spatial Empirical Bayes Map',
-    value: MappingTypes.SPATIAL_EB_MAP
-  }
-];
-
-// create an array from 2 to 20
-const defaultBins = Array.from({length: 19}, (_, i) => i + 2).map(i => ({
-  label: `${i}`,
-  value: i
-}));
 
 export function MappingPanel() {
   const intl = useIntl();
@@ -102,25 +38,6 @@ export function MappingPanel() {
     dispatchKepler,
     keplerOwnProps
   );
-
-  // useState for number of bins
-  const [k, setK] = useState(5);
-
-  // useState for variable name
-  const [variable, setVariable] = useState('');
-
-  // useState for mapping type
-  const [mappingType, setMappingType] = useState(MappingTypes.QUANTILE);
-
-  // useState for color ranges
-  const [colorRanges, setColorRanges] = useState(
-    COLOR_RANGES.filter(colorRange => {
-      return colorRange.colors.length === k;
-    })
-  );
-
-  // useState for selected color range
-  const [selectedColorRange, setSelectedColorRange] = useState(colorRanges[0]);
 
   // use selector to get tableName from redux store
   const tableName = useSelector((state: GeoDaState) => state.root.file?.rawFileData?.fileName);
@@ -137,39 +54,22 @@ export function MappingPanel() {
   // get mapStyle from redux store
   const mapStyle = useSelector((state: GeoDaState) => state.keplerGl[MAP_ID]?.mapStyle);
 
-  // get numeric columns from redux store
-  const numericColumns = useMemo(() => {
-    const fieldNames = getNumericFieldNames(layer);
-    return fieldNames;
-  }, [layer]);
+  // handle classification config changes
+  const [k, setK] = useState(5);
+  const [variable, setVariable] = useState('');
+  const [mappingType, setMappingType] = useState(MappingTypes.QUANTILE);
+  const [selectedColorRange, setSelectedColorRange] = useState(getDefaultColorRange(k));
 
-  // handle map type change
-  const onMapTypeChange = (value: any) => {
-    const selectValue = value.currentKey;
-    setMappingType(selectValue);
-  };
-
-  // handle number of bins change
-  const onKSelectionChange = (value: any) => {
-    const kValue = Number(value.currentKey);
-    // convert string to number
-    setK(kValue);
-    // update color range
-    const newColorRanges = COLOR_RANGES.filter(colorRange => {
-      return colorRange.colors.length === kValue;
-    });
-    setColorRanges(newColorRanges);
-  };
-
-  // handle variable change
-  const onVariableSelectionChange = (value: any) => {
-    const selectValue = value.currentKey;
-    setVariable(selectValue);
-  };
-
-  // handle color range selection change
-  const onSelectColorRange = (p: ColorRange) => {
-    setSelectedColorRange(p);
+  const onClassficationValueChange = ({
+    method,
+    variable,
+    k,
+    colorRange
+  }: ClassificationOnValuesChange) => {
+    setMappingType(method);
+    setVariable(variable);
+    setK(k);
+    setSelectedColorRange(colorRange);
   };
 
   // handle onCreateMap
@@ -231,71 +131,7 @@ export function MappingPanel() {
               <Card>
                 <CardBody>
                   <div className="flex flex-col gap-2">
-                    <Select
-                      label="Select map type"
-                      className="max-w"
-                      onSelectionChange={onMapTypeChange}
-                      defaultSelectedKeys={[mappingType]}
-                    >
-                      {mappingTypes.map(mappingType => (
-                        <SelectItem key={mappingType.value} value={mappingType.value}>
-                          {mappingType.label}
-                        </SelectItem>
-                      ))}
-                    </Select>
-                    <Select
-                      label="Select number of bins"
-                      className="max-w"
-                      onSelectionChange={onKSelectionChange}
-                      defaultSelectedKeys={[`${k}`]}
-                    >
-                      {defaultBins.map(bin => (
-                        <SelectItem key={bin.value} value={bin.value}>
-                          {bin.label}
-                        </SelectItem>
-                      ))}
-                    </Select>
-                    <Select
-                      label="Select variable"
-                      className="max-w"
-                      onSelectionChange={onVariableSelectionChange}
-                    >
-                      {numericColumns.map((col: string) => (
-                        <SelectItem key={col} value={col}>
-                          {col}
-                        </SelectItem>
-                      ))}
-                    </Select>
-                    <Select
-                      label="Select color scheme"
-                      className="max-w"
-                      items={colorRanges}
-                      renderValue={items => {
-                        return items.map(item => (
-                          <ColorPalette
-                            key={item.data?.name}
-                            colors={item.data?.colors || []}
-                            isReversed={false}
-                            isSelected={item.data?.name === selectedColorRange.name}
-                          />
-                        ));
-                      }}
-                      defaultSelectedKeys={[`${selectedColorRange.name}`]}
-                    >
-                      {colorRange => (
-                        <SelectItem
-                          key={`${colorRange.name}`}
-                          onClick={() => onSelectColorRange(colorRange)}
-                          className="gap-0 py-0"
-                        >
-                          <ColorPalette
-                            colors={colorRange.colors}
-                            isReversed={false}
-                            isSelected={colorRange.name === selectedColorRange.name}
-                          />
-                        </SelectItem>
-                      )}
-                    </Select>
+                    <ClassificationPanel onValuesChange={onClassficationValueChange} />
                     <Spacer y={2} />
                     <Button
                       radius="sm"
